[Tests] Split reduction tests
