+++ tests/test_general_reduction_ops.py
@@ -1,0 +1,250 @@
+import pytest
+import torch
+
+import flag_gems
+
+from .accuracy_utils import (
+    FLOAT_DTYPES,
+    REDUCTION_SHAPES,
+    REDUCTION_SMALL_SHAPES,
+    gems_assert_close,
+    gems_assert_equal,
+    skip_expr,
+    skip_reason,
+    to_reference,
+)
+from .conftest import TO_CPU
+
+FLOAT_DTYPES = [torch.float32] if TO_CPU else FLOAT_DTYPES
+DIM_LIST = [1] if TO_CPU else [0, 1]
+DIMS_LIST = [1] if TO_CPU else [0, 1, [0, 1], [1, 0]]
+KIND_KEEPDIM_DIMS_SHAPE = (
+    [("normal", True, DIMS_LIST[0], REDUCTION_SHAPES[0])]
+    if TO_CPU
+    else list(
+        zip(
+            ["normal", "allTrue"] * 2,
+            [True, False] * 2,
+            DIMS_LIST,
+            REDUCTION_SHAPES + [(7, 4, 11, 1)],
+        )
+    )
+)
+KEEPDIM_DIMS = (
+    [(True, DIMS_LIST[0])] if TO_CPU else list(zip([True, False] * 2, DIMS_LIST))
+)
+KEEPDIM_DIM = [(True, DIM_LIST[0])] if TO_CPU else list(zip([True, False], DIM_LIST))
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
+@pytest.mark.parametrize("kind", ["normal", "allTrue"])
+def test_accuracy_all_without_dim(shape, dtype, kind):
+    if kind == "allTrue":
+        inp = torch.ones(shape, dtype=dtype, device="cuda")
+    else:
+        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.all(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.all(inp)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.skipif(skip_expr, reason=skip_reason)
+@pytest.mark.parametrize("kind, keepdim, dim, shape", KIND_KEEPDIM_DIMS_SHAPE)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
+def test_accuracy_all_dims(shape, dim, keepdim, dtype, kind):
+    if kind == "allTrue":
+        inp = torch.ones(shape, dtype=dtype, device="cuda")
+    else:
+        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.all(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.all(inp, dim=dim, keepdim=keepdim)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
+@pytest.mark.parametrize("kind", ["normal", "allFalse"])
+def test_accuracy_any_without_dim(shape, dtype, kind):
+    if kind == "allFalse":
+        inp = torch.zeros(shape, dtype=dtype, device="cuda")
+    else:
+        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.any(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.any(inp)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.skipif(skip_expr, reason=skip_reason)
+@pytest.mark.parametrize("kind, keepdim, dim, shape", KIND_KEEPDIM_DIMS_SHAPE)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
+def test_accuracy_any_dims(shape, dim, keepdim, dtype, kind):
+    if kind == "allFalse":
+        inp = torch.zeros(shape, dtype=dtype, device="cuda")
+    else:
+        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.any(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.any(inp, dim=dim, keepdim=keepdim)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_max_without_dim(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.max(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.max(inp)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
+@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
+@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_max_dim(shape, dim, keepdim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out_value, ref_out_index = torch.max(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out_value, res_out_index = torch.max(inp, dim=dim, keepdim=keepdim)
+
+    gems_assert_equal(res_out_index, ref_out_index)
+    gems_assert_equal(res_out_value, ref_out_value)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_mean_without_dim(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.mean(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.mean(inp)
+
+    gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIMS)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_mean_dim(shape, dim, keepdim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.mean(ref_inp, dim, keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.mean(inp, dim, keepdim)
+
+    gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_min_without_dim(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out = torch.min(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.min(inp)
+
+    gems_assert_equal(res_out, ref_out)
+
+
+# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
+@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
+@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_min_dim(shape, dim, keepdim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp)
+
+    ref_out_value, ref_out_index = torch.min(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out_value, res_out_index = torch.min(inp, dim=dim, keepdim=keepdim)
+
+    gems_assert_equal(res_out_index, ref_out_index)
+    gems_assert_equal(res_out_value, ref_out_value)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_prod_without_dim(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.prod(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.prod(inp)
+
+    gems_assert_close(res_out, ref_out, dtype)
+
+
+# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
+@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
+@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_prod_dim(shape, dim, keepdim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.prod(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.prod(inp, dim=dim, keepdim=keepdim)
+
+    gems_assert_close(res_out, ref_out, dtype)
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_sum_without_dim(shape, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.sum(ref_inp)
+    with flag_gems.use_gems():
+        res_out = torch.sum(inp)
+
+    gems_assert_close(res_out, ref_out, dtype, reduce_dim=inp.numel())
+
+
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_sum_dim(shape, dim, keepdim, dtype):
+    inp = torch.randn(shape, dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+
+    ref_out = torch.sum(ref_inp, dim=dim, keepdim=keepdim)
+    with flag_gems.use_gems():
+        res_out = torch.sum(inp, dim=dim, keepdim=keepdim)
+
+    if isinstance(dim, int):
+        dim = [dim]
+    dim = [d % inp.ndim for d in dim]
+    _dim = 1
+    for d in dim:
+        _dim *= shape[d]
+    gems_assert_close(res_out, ref_out, dtype, reduce_dim=_dim)
