+++ tests/test_reduction_ops.py
@@ -10,8 +10,6 @@ from .accuracy_utils import (
     REDUCTION_SMALL_SHAPES,
     gems_assert_close,
     gems_assert_equal,
-    skip_expr,
-    skip_reason,
     to_reference,
 )
 from .conftest import TO_CPU
@@ -19,27 +17,11 @@ from .conftest import TO_CPU
 FLOAT_DTYPES = [torch.float32] if TO_CPU else FLOAT_DTYPES
 DIM_LIST = [1] if TO_CPU else [0, 1]
 DIMS_LIST = [1] if TO_CPU else [0, 1, [0, 1], [1, 0]]
-KIND_KEEPDIM_DIMS_SHAPE = (
-    [("normal", True, DIMS_LIST[0], REDUCTION_SHAPES[0])]
-    if TO_CPU
-    else list(
-        zip(
-            ["normal", "allTrue"] * 2,
-            [True, False] * 2,
-            DIMS_LIST,
-            REDUCTION_SHAPES + [(7, 4, 11, 1)],
-        )
-    )
-)
 KEEPDIM_DIMS_SHAPE = (
     [(True, DIMS_LIST[0], REDUCTION_SHAPES[0])]
     if TO_CPU
     else list(zip([True, False] * 2, DIMS_LIST, REDUCTION_SHAPES + [(7, 4, 11, 1)]))
 )
-KEEPDIM_DIMS = (
-    [(True, DIMS_LIST[0])] if TO_CPU else list(zip([True, False] * 2, DIMS_LIST))
-)
-KEEPDIM_DIM = [(True, DIM_LIST[0])] if TO_CPU else list(zip([True, False], DIM_LIST))
 SMOOTH_IGNORE_SHAPE = (
     [(0.1, 1, REDUCTION_SHAPES[0])]
     if TO_CPU
@@ -61,40 +43,6 @@ THRESHOLD_SHAPE = (
 CROSS_ENTROPY_LOSS_REDUCTION = ["sum"] if TO_CPU else ["mean", "none", "sum"]
 
 
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
-@pytest.mark.parametrize("kind", ["normal", "allTrue"])
-def test_accuracy_all_without_dim(shape, dtype, kind):
-    if kind == "allTrue":
-        inp = torch.ones(shape, dtype=dtype, device="cuda")
-    else:
-        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.all(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.all(inp)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-@pytest.mark.skipif(skip_expr, reason=skip_reason)
-@pytest.mark.parametrize("kind, keepdim, dim, shape", KIND_KEEPDIM_DIMS_SHAPE)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
-def test_accuracy_all_dims(shape, dim, keepdim, dtype, kind):
-    if kind == "allTrue":
-        inp = torch.ones(shape, dtype=dtype, device="cuda")
-    else:
-        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.all(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.all(inp, dim=dim, keepdim=keepdim)
-
-    gems_assert_equal(res_out, ref_out)
-
-
 @pytest.mark.parametrize("keepdim, dim, shape", KEEPDIM_DIMS_SHAPE)
 @pytest.mark.parametrize("dtype", FLOAT_DTYPES)
 def test_accuracy_amax(shape, dim, keepdim, dtype):
@@ -104,40 +52,6 @@ def test_accuracy_amax(shape, dim, keepdim, dtype):
     ref_out = torch.amax(ref_inp, dim=dim, keepdim=keepdim)
     with flag_gems.use_gems():
         res_out = torch.amax(inp, dim=dim, keepdim=keepdim)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
-@pytest.mark.parametrize("kind", ["normal", "allFalse"])
-def test_accuracy_any_without_dim(shape, dtype, kind):
-    if kind == "allFalse":
-        inp = torch.zeros(shape, dtype=dtype, device="cuda")
-    else:
-        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.any(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.any(inp)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-@pytest.mark.skipif(skip_expr, reason=skip_reason)
-@pytest.mark.parametrize("kind, keepdim, dim, shape", KIND_KEEPDIM_DIMS_SHAPE)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES + [torch.bool])
-def test_accuracy_any_dims(shape, dim, keepdim, dtype, kind):
-    if kind == "allFalse":
-        inp = torch.zeros(shape, dtype=dtype, device="cuda")
-    else:
-        inp = torch.randint(0, 2, shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.any(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.any(inp, dim=dim, keepdim=keepdim)
 
     gems_assert_equal(res_out, ref_out)
 
@@ -277,110 +191,6 @@ def test_accuracy_nonzero(shape, dtype):
     gems_assert_equal(res_out, ref_out)
 
 
-@pytest.mark.parametrize(
-    "N, C, H, W, num_groups",
-    [
-        (16, 3, 16, 16, 1),
-        (32, 32, 32, 32, 8),
-        (1, 32, 32, 32, 8),
-        (1, 32, 32, 32, 16),
-        (1, 64, 32, 32, 16),
-        (1, 64, 32, 32, 32),
-        (1, 64, 32, 32, 64),
-    ],
-)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_groupnorm(N, C, H, W, num_groups, dtype):
-    HW = H * W
-    inp = torch.randn(size=(N, C, H, W), dtype=dtype, device="cuda", requires_grad=True)
-    weight = torch.randn(size=(C,), dtype=dtype, device="cuda", requires_grad=True)
-    bias = torch.randn(size=(C,), dtype=dtype, device="cuda", requires_grad=True)
-    eps = 1e-5
-
-    ref_inp = to_reference(inp, True)
-    ref_weight = to_reference(weight, True)
-    ref_bias = to_reference(bias, True)
-
-    ref_out = torch.nn.functional.group_norm(
-        ref_inp, num_groups, weight=ref_weight, bias=ref_bias, eps=eps
-    )
-    ref_mean = torch.mean(ref_inp.reshape([N, num_groups, -1]), dim=2)
-    ref_var = torch.var(ref_inp.reshape([N, num_groups, -1]), dim=2, correction=0)
-    ref_rstd = torch.rsqrt(ref_var + eps)
-
-    (res_out, res_mean, res_rstd) = flag_gems.group_norm(
-        inp, weight, bias, N, C, HW, num_groups, eps
-    )
-
-    gems_assert_close(res_mean, ref_mean, dtype)
-    gems_assert_close(res_rstd, ref_rstd, dtype)
-    gems_assert_close(res_out, ref_out, dtype)
-
-    out_grad = torch.randn_like(inp)
-    ref_grad = to_reference(out_grad, True)
-
-    (ref_in_grad, ref_weight_grad, ref_bias_grad) = torch.autograd.grad(
-        ref_out, (ref_inp, ref_weight, ref_bias), ref_grad
-    )
-    (res_in_grad, res_weight_grad, res_bias_grad) = torch.autograd.grad(
-        res_out, (inp, weight, bias), out_grad
-    )
-    group_size = C // num_groups
-    gems_assert_close(res_in_grad, ref_in_grad, dtype, reduce_dim=group_size * HW)
-    gems_assert_close(res_weight_grad, ref_weight_grad, dtype, reduce_dim=N * HW)
-    gems_assert_close(res_bias_grad, ref_bias_grad, dtype, reduce_dim=N * HW)
-
-
-# TODO: failed at (1, 2) (2~32, 40499) (200, 2~64) (200~4096, 40999)
-@pytest.mark.parametrize("shape", [(1, 40999)] if TO_CPU else [(1, 40999), (4096, 256)])
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_layernorm(shape, dtype):
-    M = shape[0]
-    N = shape[1]
-    layer_shape = [
-        N,
-    ]
-    inp = torch.randn(shape[:2], dtype=dtype, device="cuda", requires_grad=True)
-    weight = torch.randn(layer_shape, dtype=dtype, device="cuda", requires_grad=True)
-    bias = torch.randn(layer_shape, dtype=dtype, device="cuda", requires_grad=True)
-    eps = 1e-5
-
-    ref_inp = to_reference(inp, True)
-    ref_weight = to_reference(weight, True)
-    ref_bias = to_reference(bias, True)
-
-    ref_out = torch.layer_norm(
-        ref_inp,
-        list(layer_shape),
-        weight=ref_weight,
-        bias=ref_bias,
-        eps=eps,
-    )
-    (res_out, res_mean, res_rstd) = flag_gems.layer_norm(
-        inp, list(layer_shape), weight=weight, bias=bias, eps=eps
-    )
-
-    ref_mean = torch.mean(ref_inp, dim=1)
-    ref_var = torch.var(ref_inp, dim=1, correction=0)
-    ref_rstd = torch.rsqrt(ref_var + eps)
-    gems_assert_close(res_mean, ref_mean, dtype)
-    gems_assert_close(res_rstd, ref_rstd, dtype)
-    gems_assert_close(res_out, ref_out, dtype)
-
-    out_grad = torch.randn_like(inp)
-    ref_grad = to_reference(out_grad, True)
-
-    (ref_in_grad, ref_weight_grad, ref_bias_grad) = torch.autograd.grad(
-        ref_out, (ref_inp, ref_weight, ref_bias), ref_grad
-    )
-    (res_in_grad, res_weight_grad, res_bias_grad) = torch.autograd.grad(
-        res_out, (inp, weight, bias), out_grad
-    )
-    gems_assert_close(res_in_grad, ref_in_grad, dtype, reduce_dim=N)
-    gems_assert_close(res_weight_grad, ref_weight_grad, dtype, reduce_dim=M)
-    gems_assert_close(res_bias_grad, ref_bias_grad, dtype, reduce_dim=M)
-
-
 @pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
 @pytest.mark.parametrize("dtype", FLOAT_DTYPES)
 def test_accuracy_log_softmax(shape, dtype):
@@ -399,213 +209,6 @@ def test_accuracy_log_softmax(shape, dtype):
     (ref_in_grad,) = torch.autograd.grad(ref_out, ref_inp, ref_grad)
     (res_in_grad,) = torch.autograd.grad(res_out, inp, out_grad)
     gems_assert_close(res_in_grad, ref_in_grad, dtype, reduce_dim=shape[dim])
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_max_without_dim(shape, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.max(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.max(inp)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
-@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_max_dim(shape, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out_value, ref_out_index = torch.max(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out_value, res_out_index = torch.max(inp, dim=dim, keepdim=keepdim)
-
-    gems_assert_equal(res_out_index, ref_out_index)
-    gems_assert_equal(res_out_value, ref_out_value)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_mean_without_dim(shape, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.mean(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.mean(inp)
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIMS)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_mean_dim(shape, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.mean(ref_inp, dim, keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.mean(inp, dim, keepdim)
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_min_without_dim(shape, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out = torch.min(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.min(inp)
-
-    gems_assert_equal(res_out, ref_out)
-
-
-# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
-@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_min_dim(shape, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp)
-
-    ref_out_value, ref_out_index = torch.min(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out_value, res_out_index = torch.min(inp, dim=dim, keepdim=keepdim)
-
-    gems_assert_equal(res_out_index, ref_out_index)
-    gems_assert_equal(res_out_value, ref_out_value)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_prod_without_dim(shape, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.prod(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.prod(inp)
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-# TODO: failed at (200, 40999, 3), while successed at this shape in mean_dim
-@pytest.mark.parametrize("shape", REDUCTION_SMALL_SHAPES)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_prod_dim(shape, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.prod(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.prod(inp, dim=dim, keepdim=keepdim)
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_rmsnorm(shape, dtype):
-    N = shape[1]
-    layer_shape = [
-        N,
-    ]
-    inp = torch.randn(shape[:2], dtype=dtype, device="cuda")
-    weight = torch.randn(layer_shape, dtype=dtype, device="cuda")
-    eps = 1e-5
-
-    ref_inp = to_reference(inp, True)
-    ref_weight = to_reference(weight, True)
-
-    def _torch_rms_norm(x, weight, eps):
-        variance = x.pow(2).mean(-1, keepdim=True)
-        hidden_states = x * torch.rsqrt(variance + eps)
-        return weight * hidden_states
-
-    ref_out = _torch_rms_norm(ref_inp, weight=ref_weight, eps=eps)
-
-    res_out = flag_gems.rms_norm(inp, list(layer_shape), weight=weight, eps=eps)
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_skip_layernorm(shape, dtype):
-    N = shape[1]
-    layer_shape = [
-        N,
-    ]
-    inp = torch.randn(shape[:2], dtype=dtype, device="cuda")
-    residual = torch.randn(shape[:2], dtype=dtype, device="cuda")
-    weight = torch.randn(layer_shape, dtype=dtype, device="cuda")
-    bias = torch.randn(layer_shape, dtype=dtype, device="cuda")
-    eps = 1e-5
-
-    ref_inp = to_reference(inp, True)
-    ref_residual = to_reference(residual, True)
-    ref_weight = to_reference(weight, True)
-    ref_bias = to_reference(bias, True)
-
-    ref_out = torch.layer_norm(
-        ref_inp + ref_residual,
-        list(layer_shape),
-        weight=ref_weight,
-        bias=ref_bias,
-        eps=eps,
-    )
-    res_out = flag_gems.skip_layer_norm(
-        inp, residual, list(layer_shape), weight=weight, bias=bias, eps=eps
-    )
-
-    gems_assert_close(res_out, ref_out, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_skip_rmsnorm(shape, dtype):
-    N = shape[1]
-    layer_shape = [
-        N,
-    ]
-    inp = torch.randn(shape[:2], dtype=dtype, device="cuda")
-    residual = torch.randn(shape[:2], dtype=dtype, device="cuda")
-    weight = torch.randn(layer_shape, dtype=dtype, device="cuda")
-    eps = 1e-5
-
-    ref_inp = to_reference(inp, True)
-    ref_residual = to_reference(residual, True)
-    ref_weight = to_reference(weight, True)
-
-    def _torch_rms_norm(x, residual, weight, eps):
-        x = x + residual
-        variance = x.pow(2).mean(-1, keepdim=True)
-        hidden_states = x * torch.rsqrt(variance + eps)
-        return weight * hidden_states
-
-    ref_out = _torch_rms_norm(
-        ref_inp,
-        ref_residual,
-        weight=ref_weight,
-        eps=eps,
-    )
-
-    res_out = flag_gems.skip_rms_norm(
-        inp, residual, list(layer_shape), weight=weight, eps=eps
-    )
-
-    gems_assert_close(res_out, ref_out, dtype)
 
 
 # TODO: failed at (1, 2) (200, 40999, 3)
@@ -632,39 +235,6 @@ def test_accuracy_softmax(shape, dtype, dim):
 
 
 @pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_sum_without_dim(shape, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.sum(ref_inp)
-    with flag_gems.use_gems():
-        res_out = torch.sum(inp)
-
-    gems_assert_close(res_out, ref_out, dtype, reduce_dim=inp.numel())
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIM)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_sum_dim(shape, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.sum(ref_inp, dim=dim, keepdim=keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.sum(inp, dim=dim, keepdim=keepdim)
-
-    if isinstance(dim, int):
-        dim = [dim]
-    dim = [d % inp.ndim for d in dim]
-    _dim = 1
-    for d in dim:
-        _dim *= shape[d]
-    gems_assert_close(res_out, ref_out, dtype, reduce_dim=_dim)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
 @pytest.mark.parametrize("dim", DIMS_LIST)
 @pytest.mark.parametrize("correction", [1] if TO_CPU else [0, 1])
 @pytest.mark.parametrize("keepdim", [True] if TO_CPU else [True, False])
@@ -685,23 +255,6 @@ def test_accuracy_varmean(shape, dim, correction, keepdim, dtype):
 
     gems_assert_close(res_mean, ref_mean, dtype)
     gems_assert_close(res_var, ref_var, dtype)
-
-
-@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
-@pytest.mark.parametrize(
-    "ord", [2] if TO_CPU else [2, float("inf"), -float("inf"), 0, 1]
-)
-@pytest.mark.parametrize("keepdim, dim", KEEPDIM_DIMS)
-@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
-def test_accuracy_vectornorm(shape, ord, dim, keepdim, dtype):
-    inp = torch.randn(shape, dtype=dtype, device="cuda")
-    ref_inp = to_reference(inp, True)
-
-    ref_out = torch.linalg.vector_norm(ref_inp, ord, dim, keepdim)
-    with flag_gems.use_gems():
-        res_out = torch.linalg.vector_norm(inp, ord, dim, keepdim)
-
-    gems_assert_close(res_out, ref_out, dtype)
 
 
 @pytest.mark.skip(reason="operator undone")
